Basic texture support
- Added stb_image
- Added script to build & run
- Added warning as errors into build

+++ kandinsky/defines.h
@@ -2,8 +2,8 @@
 
 #include <cstdint>
 
-using i16 = int16_t;
-using u16 = uint16_t;
+using i8 = int8_t;
+using u8 = uint8_t;
 
 using i16 = int16_t;
 using u16 = uint16_t;
